+++ MazeGenerator/Ray.h
@@ -4,5 +4,5 @@ class Ray : public Segment {
 public:
 	Ray();
 	Ray(Coordinates p1, Coordinates p2);
-	Ray(Coordinates p1, float length, double angle);
+	Ray(Coordinates p1, double length, double angle);
 };
